gpu: mali: Use reserved memory only when it is really reserved
Registering a chunk of reserved memory at a hardcoded location in the
Mali400 GPU resources table is wrong when CONFIG_SUNXI_MALI_RESERVED_MEM
option is not enabled.

After this fix, Mali400 GPU is configured to use up to 256 MiB of
normal memory when CONFIG_SUNXI_MALI_RESERVED_MEM is not enabled.
When CONFIG_SUNXI_MALI_RESERVED_MEM is enabled, it uses 64 MiB of
reserved memory and up to 192 MiB of normal memory (same as before).

Signed-off-by: Siarhei Siamashka <siarhei.siamashka@gmail.com>