Add smp support for Allwinner A20(sunxi 7i).
This patch adds SMP support for the Allwinner A20 SoC. This SoC uses an IP to, among other things, handle the CPU-related configuration, like the power clamp, the boot address of the secondary CPUS, etc. We thus need to map this IP during the prepare_cpu SMP operation, before bringing up the secondary CPU in the secondary_startup operation.

Signed-off-by: Fan Rong <cinifr@gmail.com>
Signed-off-by: Maxime Ripard <maxime.ripard@free-electrons.com>

Conflicts:
	arch/arm/mach-sunxi/Makefile
	arch/arm/mach-sunxi/headsmp.S
	arch/arm/mach-sunxi/platsmp.c
	arch/arm/mach-sunxi/sunxi.c