ARM: sunxi: Extend stmmac to support GMAC in Allwinner A20 SoC's
The Allwinner A20 has an ethernet controller that seems to be
identical to Synopsys DesignWare MAC 10/100/1000 Universal,
which is supported by the stmmac driver.

Allwinner's GMAC requires setting additional registers in the SoC's
clock control unit.

The exact version of the DWMAC IP that Allwinner uses is unknown,
thus the exact feature set is unknown.

Conflicts:
	arch/arm/mach-sunxi/sunxi.c