of: overlay: Detect broken blobs.
When error is reported trying to apply an overlay for some reason sysfs
tries to call the store function again with a 2-byte offset into the dt
blob. This blob is invalid so of_fdt_unflatten_tree does not set the
root node pointer which is used in the of_resolve_phandles call.

of_fdt_unflatten_tree now returns NULL on failure so use the return
value instead of doing dances with setting the result to NULL and
checking if it changed.

Signed-off-by: Michal Suchanek <hramrach@gmail.com>
---
 * update for of_fdt_unflatten_tree prototype change