of: overlay: Detect broken blobs.
When error is reported trying to apply an overlay for some reason sysfs
tries to call the store function again with a 2-byte offset into the dt
blob. This blob is invalid so of_fdt_unflatten_tree does not set the
root node pointer which is used in the of_resolve_phandles call.

Set the pointer to NULL beforehand and only proceed if it's non-NULL
after of_fdt_unflatten_tree.

Signed-off-by: Michal Suchanek <hramrach@gmail.com>