mtd: partitions: remove kmemdup()
The use of kmemdup() complicates the error handling a bit. We don't
actually need to allocate new memory, since this reference is treated as
const, and it is copied into new memory by the partition registration
code anyway. So remove it.

NOTE: there's a warning in parse_mtd_partitions() about dropping the
'const'; this can be fixed... later

Signed-off-by: Brian Norris <computersforpeace@gmail.com>