usb: select NOP_USB_XCEIV by drivers that require it
NOP_USB_XCEIV is non-obvious dependency for MUSB and other drivers.

This is a virtual driver in the sense that there is no actual piece of
hardware you can point at and say you did not include driver for this so
it won't work.

Some defconfigs have it included and some which seem in need of it
don't.

So just change all depends on it to select.

Signed-off-by: Michal Suchanek <hramrach@gmail.com>