spi: sun4i: set maximum and minimum speed of SPI master.
The maximum speed of SPI master is used when maximum speed of SPI slave
is not specified. Also the __spi_validate function should check that
transfer speed do not exceed the master limits.

The user manual for A10 and A31 specifies maximum
speed of the SPI clock as 100MHz and minimum as 3kHz.

Signed-off-by: Michal Suchanek <hramrach@gmail.com>